requirements.txt
numpy
matplotlib
pandas
torch
ucimlrepo
scikit-learn
keras
tensorflow
opencv-python-headless
Pillow
pydot==1.4.2